app/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-filter==23.3
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
phonenumbers==8.13.25
Pillow==10.1.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
